Bugfix Fehler beim speichern zukünftiger Beitragsgruppen https://sourceforge.net/p/jverein/tickets/228/

+++ src/de/jost_net/JVerein/server/MitgliedNextBGruppeImpl.java
@@ -62,7 +62,7 @@ public class MitgliedNextBGruppeImpl extends AbstractDBObject implements
   {
     if (null == mitglied)
       throw new RemoteException("Mitglied fehlt!!");
-    setAttribute(COL_MITGLIED, new Integer(mitglied.getID()));
+    setAttribute(COL_MITGLIED, new Long(mitglied.getID()));
   }
 
   @Override
@@ -82,7 +82,7 @@ public class MitgliedNextBGruppeImpl extends AbstractDBObject implements
   {
     if (null == beitragsGruppe)
       throw new RemoteException("Beitragsgruppe fehlt!!");
-    setAttribute(COL_BEITRAGSGRUPPE, new Integer(beitragsGruppe.getID()));
+    setAttribute(COL_BEITRAGSGRUPPE, new Long(beitragsGruppe.getID()));
   }
 
   @Override
